Add in suggestion from Matt Welland and decrease n workers
As per Matt Wellands suggestion, I updated filtering the `mt.filters` to handle cases where filters can be missing or a 0 length set. I also decreased the number of secondary workers to the deafult (the 20 workers was a holdover from teh last script) and I exported data to parquet, which will make reading in the genotype file much quicker.